efficientNetDev/requirements.txt
torch==2.6.0+cpu
torchvision==0.21.0+cpu
matplotlib==3.8.0
pandas==2.1.4
Pillow==10.2.0
tqdm==4.65.0
